storage: breaking changes for rkyv support with u64
Makes `SegmentIndex`, `MaxCut`, and `PolicyId` opaque wrappers of `u64`.

This will allow them to be `u64_le` wrappers later so they can implement `rkyv::Portable` and thus archive as themselves.

+++ crates/aranya-runtime/src/storage/linear/io.rs
@@ -39,7 +39,7 @@ pub trait Write {
     /// A function is used to allow the item to contain its offset.
     fn append<F, T>(&mut self, builder: F) -> Result<T, StorageError>
     where
-        F: FnOnce(usize) -> T,
+        F: FnOnce(u64) -> T,
         T: Serialize;
 
     /// Set the commit head.
@@ -49,7 +49,7 @@ pub trait Write {
 /// A share-able reader for a linear storage graph.
 pub trait Read: Clone {
     /// Fetch an item written by `Write::append`.
-    fn fetch<T>(&self, offset: usize) -> Result<T, StorageError>
+    fn fetch<T>(&self, offset: u64) -> Result<T, StorageError>
     where
         T: DeserializeOwned;
 }
